[Refactor] Centralize balance display and betting logic in main loop
Refactored the game to centralize the balance display and betting logic within the main loop of `pygame_screen.py`. The `Inside` and `Outside` classes now reference the main screen instance to access and update the balance.

- Removed redundant balance and message display methods from `Inside` and `Outside` classes.
- Updated constructors of `Inside` and `Outside` to accept the main game screen instance.
- Ensured balance is updated correctly after each bet.
- Fixed the balance display to clear the previous balance before updating.
- Simplified the main loop to ensure `display_balance` is called consistently.

This change ensures that the current balance is always displayed correctly and avoids redundancy in balance update logic.
