+++ outside.py
@@ -6,32 +6,23 @@ from color import Color
 class Outside:
     """Class to Handle Outside Bets"""
 
-    def __init__(self, screen, roulette, font, background_color, accent_color, result_color, clock, spin_wheel_animation):
-        self.screen = screen
-        self.roulette = roulette
-        self.font = font
-        self.background_color = background_color
-        self.accent_color = accent_color
-        self.result_color = result_color
-        self.balance = 100
-        self.clock = clock
-        self.spin_wheel_animation = spin_wheel_animation
-
-    def display_message(self, message, pos, color=None):
-        """Render and display a message on the screen."""
-        if color is None:
-            color = self.accent_color
-        text = self.font.render(message, True, color)  # Render the text to an image
-        self.screen.blit(text, pos)  # Draw the text image on the screen at the specified position
-
-    def display_balance(self):
-        """Display the current balance at the top of the screen."""
-        self.display_message(f"Current Balance: ${self.balance}", (100, 50), self.accent_color)
+    def __init__(self, game_screen):
+        self.screen = game_screen.screen
+        self.roulette = game_screen.roulette
+        self.font = game_screen.font
+        self.background_color = game_screen.background_color
+        self.accent_color = game_screen.accent_color
+        self.result_color = game_screen.result_color
+        self.clock = game_screen.clock
+        self.spin_wheel_animation = game_screen.spin_wheel_animation
+        self.display_message = game_screen.display_message
+        self.get_user_input = game_screen.get_user_input
+        self.game_screen = game_screen
 
     def display_outside_bets_menu(self):
         """Display the outside bets menu."""
         self.screen.fill(self.background_color)  # Fill the screen with the background color
-        self.display_balance()
+        self.game_screen.display_balance()
         self.display_message('Outside Bets', (100, 100), self.accent_color)
         self.display_message('1. Bet on a color', (100, 200))
         self.display_message('2. Bet on odd or even', (100, 300))
@@ -55,54 +46,6 @@ class Outside:
             except ValueError:
                 self.display_message("Invalid input. Please enter a number.", (100, 650))
 
-    def get_user_input(self, prompt="", menu_function=None):
-        """Get user input from the screen."""
-        input_box = pygame.Rect(300, 650, 140, 48)  # Create a rectangle for the input box
-        color_inactive = pygame.Color('lightskyblue3')  # Inactive color for the input box
-        color_active = self.accent_color  # Active color for the input box
-        color = color_active  # Start with the active color to automatically focus
-        active = True  # Automatically set the input box to active
-        text = ''  # Text entered by the user
-        done = False  # State to track when input is complete
-
-        while not done:
-            for event in pygame.event.get():
-                if event.type == pygame.QUIT:  # Handle window close button
-                    pygame.quit()
-                    sys.exit()
-                if event.type == pygame.MOUSEBUTTONDOWN:  # Handle mouse click
-                    if input_box.collidepoint(event.pos):
-                        active = not active
-                    else:
-                        active = False
-                    color = color_active if active else color_inactive
-                if event.type == pygame.KEYDOWN:  # Handle key press
-                    if active:
-                        if event.key in [pygame.K_RETURN, pygame.K_KP_ENTER]:  # Complete input on Enter or keypad Enter key
-                            done = True
-                        elif event.key == pygame.K_BACKSPACE:  # Remove last character on Backspace key
-                            text = text[:-1]
-                        else:  # Add typed character to the text
-                            text += event.unicode
-
-            self.screen.fill(self.background_color)  # Clear the screen with background color
-            self.display_balance()
-            if menu_function:
-                menu_function()  # Call the appropriate menu function
-            self.display_message(prompt, (100, 600))  # Display the prompt
-            txt_surface = self.font.render(text, True, color)  # Render the text entered by the user
-            width = max(200, txt_surface.get_width() + 10)  # Set the width of the input box
-            input_box.w = width  # Update the input box width
-            self.screen.blit(txt_surface, (input_box.x + 5, input_box.y + 5))  # Draw the text on the screen
-            pygame.draw.rect(self.screen, color, input_box, 2)  # Draw the input box
-            roulette_board = pygame.image.load("Roulette_Board.png")
-            roulette_board = pygame.transform.scale(roulette_board, (250, 500))
-            self.screen.blit(roulette_board, (500, 100))  # Blit the Surface object directly
-            pygame.display.flip()  # Update the full display Surface to the screen
-            self.clock.tick(30)  # Control the loop's frame rate
-
-        return text  # Return the text entered by the user
-
     def get_color_bet(self):
         """Get the user's color bet, ensuring valid input."""
         while True:
@@ -120,7 +63,6 @@ class Outside:
         self.spin_wheel_animation()
         ball, result_color = self.roulette.spin()  # Spin the wheel
         self.screen.fill(self.background_color)
-        self.display_balance()
         self.display_message(f"The ball landed in pocket {ball} ({result_color.name.lower()})", (100, 700), self.result_color)
         pygame.display.flip()
         pygame.time.wait(2000)
@@ -128,10 +70,10 @@ class Outside:
         if self.roulette.isWinnerByColor(color):  # Check if the user won
             payout_ratio = 35 if color == Color.GREEN else 1
             winnings = self.roulette.calculateWinnings(bet, payout_ratio)
-            self.balance += winnings
+            self.game_screen.balance += winnings
             self.display_message(f"You won ${winnings}!", (100, 750))
         else:
-            self.balance -= bet
+            self.game_screen.balance -= bet
             self.display_message(f"You lost ${bet}.", (100, 750))
         pygame.display.flip()
         pygame.time.wait(2000)
@@ -153,7 +95,6 @@ class Outside:
         self.spin_wheel_animation()
         ball, color = self.roulette.spin()  # Spin the wheel
         self.screen.fill(self.background_color)
-        self.display_balance()
         self.display_message(f"The ball landed in pocket {ball} ({color.name.lower()})", (100, 700), self.result_color)
         pygame.display.flip()
         pygame.time.wait(2000)
@@ -161,10 +102,10 @@ class Outside:
         if self.roulette.isWinnerByOddEven(odd_even):  # Check if the user won
             payout_ratio = 1
             winnings = self.roulette.calculateWinnings(bet, payout_ratio)
-            self.balance += winnings
+            self.game_screen.balance += winnings
             self.display_message(f"You won ${winnings}!", (100, 750))
         else:
-            self.balance -= bet
+            self.game_screen.balance -= bet
             self.display_message(f"You lost ${bet}.", (100, 750))
         pygame.display.flip()
         pygame.time.wait(2000)
@@ -186,7 +127,6 @@ class Outside:
         self.spin_wheel_animation()
         ball, color = self.roulette.spin()  # Spin the wheel
         self.screen.fill(self.background_color)
-        self.display_balance()
         self.display_message(f"The ball landed in pocket {ball} ({color.name.lower()})", (100, 700), self.result_color)
         pygame.display.flip()
         pygame.time.wait(2000)
@@ -194,10 +134,10 @@ class Outside:
         if self.roulette.isWinnerByHighLow(high_low):  # Check if the user won
             payout_ratio = 1
             winnings = self.roulette.calculateWinnings(bet, payout_ratio)
-            self.balance += winnings
+            self.game_screen.balance += winnings
             self.display_message(f"You won ${winnings}!", (100, 750))
         else:
-            self.balance -= bet
+            self.game_screen.balance -= bet
             self.display_message(f"You lost ${bet}.", (100, 750))
         pygame.display.flip()
         pygame.time.wait(2000)
